+++ test/runtests.jl
@@ -1,3 +1,15 @@
+# Pick correct test environment.
+if ["pinned"] == ARGS
+    include("../compat/to_pinned.jl")
+elseif ["lower"] == ARGS
+    include("../compat/to_lower.jl")
+elseif ["latest"] == ARGS || isempty(ARGS)
+    # Regular testing with latest compatible versions.
+else
+    error("Invalid test arguments: $ARGS")
+end
+
+import CompatHelperLocal
 using Crayons
 bold = crayon"bold"
 blue = crayon"blue"
@@ -30,3 +42,6 @@ sep("Run doctests (DEACTIVATED while migrating api from 'Internals').")
 
 sep("Check source code formatting.")
 include("./formatting.jl")
+
+sep("Check compatibility entries.")
+CompatHelperLocal.@check()